CHORE: login되었을 때 SideBar 내부에 알림 토글 버튼 생성
push notification을 구현하기 위해서는 user browser에서
permission을 설정해 주어야 하는데, 한 번 허용을 하지 않게 하면
user가 다시 알림 신청을 하기 위해서는 번거로움을 감소해야
하기 때문에 버튼을 만들어 주었다.
